[step2] H2 database 에 User 클래스 추가
* User 클래스에 @Entity 애노테이션 생성
* domain 패키지 생성
    * User 클래스를 domain 패키지 밑으로 이동
    * UserRepository 인터페이스 생성
* application.properties 수정
* build.gradle 수정
[에러처리] h2 db 때문에 문제가 생긴다면 application.properties와 build.gradle을 유심히 살펴보자
