+++ src/main/java/com/codessquad/qna/UserController.java
@@ -1,6 +1,8 @@
 package com.codessquad.qna;
 
+import com.codessquad.qna.domain.User;
 import org.springframework.stereotype.Controller;
+import org.springframework.stereotype.Service;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
