+++ src/main/java/com/codessquad/qna/domain/Question.java
@@ -1,0 +1,59 @@
+package com.codessquad.qna.domain;
+
+public class Question {
+
+    public int index;
+    public String writer;
+    public String title;
+    public String contents;
+    public String localDateTime;
+
+    public String getLocalDateTime() {
+        return localDateTime;
+    }
+
+    public void setLocalDateTime(String localDateTime) {
+        this.localDateTime = localDateTime;
+    }
+
+    public int getIndex() { return index; }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public String getContents() {
+        return contents;
+    }
+
+    public String getWriter() {
+        return writer;
+    }
+
+    public void setIndex(int index) {
+        this.index = index;
+    }
+
+    public void setWriter(String writer) {
+        this.writer = writer;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public void setContents(String contents) {
+        this.contents = contents;
+    }
+
+    @Override
+    public String toString() {
+        return "Question{" +
+                "index=" + index +
+                ", writer='" + writer + '\'' +
+                ", title='" + title + '\'' +
+                ", contents='" + contents + '\'' +
+                ", localDateTime='" + localDateTime + '\'' +
+                '}';
+    }
+}
