+++ src/main/java/com/codessquad/qna/domain/UserRepository.java
@@ -1,0 +1,6 @@
+package com.codessquad.qna.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
