+++ src/main/resources/application.properties
@@ -5,7 +5,7 @@ spring.devtools.livereload.enabled=true
 spring.devtools.restart.enabled=false
 spring.thymeleaf.cache=false
 
-spring.datasource.url=jdbc:h2:mem://localhost/~/java-qna;MVCC=TRUE;DB_CLOSE_ON_EXIT=FALSE
+spring.datasource.url=jdbc:h2:~/java-qna;MVCC=TRUE;DB_CLOSE_ON_EXIT=FALSE
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
 spring.datasource.password=
@@ -15,7 +15,7 @@ spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 
-spring.jpa.hibernate.ddl-auto=create-drop
-spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.hibernate.ddl-auto=create
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
+
